# backend/docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"UserProfile": {
				"properties": {
					"id": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"avatarUrl": {
						"type": "string"
					},
					"bio": {
						"type": "string"
					},
					"aboutMe": {
						"type": "string"
					},
					"phoneCountryCode": {
						"type": "number",
						"format": "double"
					},
					"phoneNumber": {
						"type": "number",
						"format": "double"
					},
					"yearOfBirth": {
						"type": "number",
						"format": "double"
					},
					"location": {
						"type": "string"
					},
					"userType": {
						"type": "string"
					},
					"themeColor": {
						"type": "string"
					},
					"spotifyPlaylistUrl": {
						"type": "string"
					},
					"onboardingCompleted": {
						"type": "boolean"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"username",
					"firstName",
					"lastName",
					"location",
					"userType",
					"onboardingCompleted",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"CollaborationResponse": {
				"description": "API response format for a user collaboration\r\nRepresents a past collaboration between two users",
				"properties": {
					"id": {
						"type": "string"
					},
					"userId": {
						"type": "string"
					},
					"collaboratorId": {
						"type": "string"
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"createdAt": {
						"type": "string",
						"nullable": true
					},
					"collaborator": {
						"properties": {
							"avatarUrl": {
								"type": "string",
								"nullable": true
							},
							"lastName": {
								"type": "string",
								"nullable": true
							},
							"firstName": {
								"type": "string",
								"nullable": true
							},
							"username": {
								"type": "string"
							},
							"id": {
								"type": "string"
							}
						},
						"required": [
							"username",
							"id"
						],
						"type": "object"
					}
				},
				"required": [
					"id",
					"userId",
					"collaboratorId",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"CreateCollaborationDto": {
				"description": "DTO for creating a collaboration\r\n\r\nRepresents a past collaboration between the authenticated user and another user.\r\nThe description is optional and limited to 200 characters.",
				"properties": {
					"description": {
						"type": "string",
						"nullable": true,
						"description": "Optional description of the collaboration (max 200 characters)",
						"example": "Worked together on a jazz album in 2023"
					},
					"collaborator_id": {
						"type": "string",
						"description": "The ID of the user you collaborated with (UUID format)",
						"example": "a1b2c3d4-e5f6-7890-1234-567890abcdef"
					}
				},
				"required": [
					"collaborator_id"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"collaborator_id": "a1b2c3d4-e5f6-7890-1234-567890abcdef",
					"description": "Worked together on a jazz album in 2023"
				}
			},
...
			"AuthResponse": {
				"properties": {
					"user": {
						"properties": {
							"createdAt": {
								"type": "string"
							},
							"email": {
								"type": "string"
							},
							"id": {
								"type": "string"
							}
						},
						"required": [
							"createdAt",
							"email",
							"id"
						],
						"type": "object"
					},
					"session": {
						"properties": {
							"expiresAt": {
								"type": "number",
								"format": "double"
							},
							"expiresIn": {
								"type": "number",
								"format": "double"
							},
							"refreshToken": {
								"type": "string"
							},
							"accessToken": {
								"type": "string"
							}
						},
						"required": [
							"expiresAt",
							"expiresIn",
							"refreshToken",
							"accessToken"
						],
						"type": "object"
					},
					"profile": {
						"$ref": "#/components/schemas/UserProfile"
					}
				},
				"required": [
					"user",
					"session",
					"profile"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SignupDto": {
				"description": "DTO for user signup\r\n\r\nValidates all required fields for creating a new user account.\r\nThe username, email, and phone number combination must be unique.",
				"properties": {
					"email": {
						"type": "string",
						"description": "User's email address (10-60 characters)",
						"example": "user@example.com"
					},
					"password": {
						"type": "string",
						"description": "Password (6-60 characters, must contain uppercase, lowercase, and number)",
						"example": "SecurePass123"
					},
					"username": {
						"type": "string",
						"description": "Unique username (3-30 characters, alphanumeric and underscores only)",
						"example": "johndoe"
					},
					"firstName": {
						"type": "string",
						"description": "First name (2-50 characters)",
						"example": "John"
					},
					"lastName": {
						"type": "string",
						"description": "Last name (2-50 characters)",
						"example": "Doe"
					},
					"phoneCountryCode": {
						"type": "number",
						"format": "double",
						"description": "Phone country code (1-999)",
						"example": 1
					},
					"phoneNumber": {
						"type": "number",
						"format": "double",
						"description": "Phone number (4-15 digits)",
						"example": 1234567890
					},
					"yearOfBirth": {
						"type": "number",
						"format": "double",
						"description": "Year of birth (must be 13+ years old)",
						"example": 1990
					},
					"location": {
						"type": "string",
						"description": "User's location (1-100 characters)",
						"example": "New York, NY"
					},
					"userType": {
						"type": "string",
						"enum": [
							"musician",
							"service_provider",
							"other"
						],
						"description": "User type: \"musician\", \"service_provider\", or \"other\"",
						"example": "musician"
					}
				},
				"required": [
					"email",
					"password",
					"username",
					"firstName",
					"lastName",
					"phoneCountryCode",
					"phoneNumber",
					"yearOfBirth",
					"location",
					"userType"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"email": "user@example.com",
					"password": "SecurePass123",
					"username": "johndoe",
					"firstName": "John",
					"lastName": "Doe",
					"phoneCountryCode": 1,
					"phoneNumber": 1234567890,
					"yearOfBirth": 1990,
					"location": "New York, NY",
					"userType": "musician"
				}
			},
			"LoginDto": {
				"description": "DTO for user login\r\n\r\nUsed to authenticate an existing user with their email and password.",
				"properties": {
					"email": {
						"type": "string",
						"description": "User's email address",
						"example": "user@example.com"
					},
					"password": {
						"type": "string",
						"description": "User's password",
						"example": "SecurePass123"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"email": "user@example.com",
					"password": "SecurePass123"
				}
			}
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "LineUp API",
		"version": "1.0.0",
		"description": "API documentation for the LineUp application",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Pinnaacle"
		}
	},
	"paths": {
...
		"/collaborations": {
			"get": {
				"operationId": "GetCollaborations",
				"responses": {
					"200": {
						"description": "Array of collaborations with collaborator details",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/CollaborationResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get all collaborations for the authenticated user\r\n\r\nReturns a list of all collaborations where the authenticated user is involved.\r\nEach collaboration includes the collaborator's profile information.",
				"summary": "Get user's collaborations",
				"tags": [
					"Collaborations"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "CreateCollaboration",
				"responses": {
					"200": {
						"description": "The created collaboration",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CollaborationResponse"
								}
							}
						}
					}
				},
				"description": "Create a collaboration\r\n\r\nAllows the authenticated user to record a past collaboration with another user.\r\nThe collaborator must exist and the collaboration must not already exist.",
				"summary": "Create a collaboration",
				"tags": [
					"Collaborations"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "Collaboration data including collaborator ID and optional description",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateCollaborationDto",
								"description": "Collaboration data including collaborator ID and optional description"
							}
						}
					}
				}
			}
		},
		"/collaborations/{userId}": {
			"get": {
				"operationId": "GetUserCollaborations",
				"responses": {
					"200": {
						"description": "Array of collaborations with collaborator details",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/CollaborationResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get collaborations for a specific user\r\n\r\nReturns all collaborations for the specified user. This endpoint can be\r\naccessed without authentication to view public collaboration information.",
				"summary": "Get collaborations by user ID",
				"tags": [
					"Collaborations"
				],
				"security": [],
				"parameters": [
					{
						"description": "The UUID of the user to retrieve collaborations for",
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/collaborations/{collaborationId}": {
			"delete": {
				"operationId": "DeleteCollaboration",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Delete a collaboration\r\n\r\nRemoves a collaboration. Only the user who created the collaboration\r\ncan delete it.",
				"summary": "Delete a collaboration",
				"tags": [
					"Collaborations"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"description": "The UUID of the collaboration to delete",
						"in": "path",
						"name": "collaborationId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/auth/signup": {
			"post": {
				"operationId": "Signup",
				"responses": {
					"200": {
						"description": "Authentication response with user, session tokens, and profile",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthResponse"
								}
							}
						}
					}
				},
				"description": "Sign up a new user\r\n\r\nCreates a new user account with email, password, and basic profile information.\r\nThe username, email, and phone number must be unique. After successful signup,\r\nreturns authentication tokens and user profile.",
				"summary": "Create a new user account",
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "User registration data including email, password, username, and profile info",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SignupDto",
								"description": "User registration data including email, password, username, and profile info"
							}
						}
					}
				}
			}
		},
		"/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Authentication response with user, session tokens, and profile",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthResponse"
								}
							}
						}
					}
				},
				"description": "Sign in an existing user\r\n\r\nAuthenticates a user with their email and password. Returns authentication\r\ntokens (access and refresh) along with user profile information.",
				"summary": "Authenticate user",
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "Login credentials (email and password)",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginDto",
								"description": "Login credentials (email and password)"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}